pyproject.toml
[tool.poetry]
name = "week2"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "src"}]


[tool.poetry.dependencies]
python = "3.11.6"
openai = "^1.3.7"
langchain = "^0.0.346"
pypdf = "^3.17.1"
cohere = "^4.37"
tiktoken = "^0.5.2"
pinecone-client = {extras = ["grpc"], version = "^2.2.4"}
chromadb = "^0.4.18"
wandb = "^0.16.1"
chainlit = "^0.7.700"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.0"
black = "^23.11.0"
isort = "^5.12.0"
ipykernel = "^6.27.1"


[tool.poetry.group.data.dependencies]
arxiv = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
